+++ safespaces/devices/desktop.lua
@@ -1,0 +1,8 @@
+--
+-- basic "windowed 3D WM"
+--
+return {
+	disable_vrbridge = true,
+	bindings = {
+	}
+};