+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 0.2 (ongoing)
 
 * Added 'prelaunch' option to device profile
-* Added Rift device profile
-* Added Vive device profile
+* New device profiles for: Rift, Vive, LG100
+* Der-B added controls for handling rotated displays
+
+New controls / options:
+* /mouse=ipd : adjust HMD parameters (cycle for distortion, abberatio, etc.)
+
 * Added menu controls for grab toggle
-* Added controls for handling rotated displays (Thanks-to: der-b)
+
 
 0.1 : 28/3/2018 - First Release